remove old test, add couterValue check
